mejoras visuales

+++ src/app/auth/administrador/administrador.component.ts
@@ -4,9 +4,9 @@ import Swal from 'sweetalert2';
 import { Usuario } from 'src/app/models/user.model';
 import { AdminService } from 'src/app/services/admin.service';
 import { MatDialog } from '@angular/material/dialog';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { DialogComponent } from './dialog/dialog.component';
 import { MatTableDataSource } from '@angular/material/table';
+import { ToastrService } from 'ngx-toastr';
 
 
 @Component({
@@ -18,18 +18,21 @@ import { MatTableDataSource } from '@angular/material/table';
 export class AdministradorComponent implements OnInit {
 
   public loading: boolean // mostrar el icono de cargar
-  constructor(private adminSrv: AdminService, private dialog: MatDialog, private snack: MatSnackBar, ){
+  constructor(private adminSrv: AdminService, private dialog: MatDialog, private toast: ToastrService, ){
     this.loading = true
   }
 
   usuarios: Usuario[] = [] // Todos los usuarios
   dataSource = new MatTableDataSource(this.usuarios);
+  loggedUser: string = '' //Nombre del usuario logeado
   
   displayedColumns: string[] = ['Admin', 'Nombre', 'Nombre de Usuario', 'Fec. Creación', 'Estado','Bloqueo', 'Contraseña', 'Editar', 'Eliminar'];
   
   ngOnInit(): void {
     this.cargarUsuarios()
     this.hideLoader()
+
+    this.loggedUser = localStorage.getItem('usuario')?.replace(/"+/g,'') || ''
   }
   
   applyFilter(event: Event) {
@@ -44,31 +47,51 @@ export class AdministradorComponent implements OnInit {
       this.dataSource = new MatTableDataSource(this.usuarios);
     }, (err)=> {
       console.warn(err) 
-      this.snack.open(err.error.msg, 'Error', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+      this.toast.error(err.error.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
     })
   }
 
-  async deleteUser(id: string, userName: string){
+  async deleteUser(usuario: Usuario){
+
+    if (usuario.tipo == 0 && this.loggedUser != 'ADMIN') {
+      this.toast.error('Un Administrador no puede BORRAR a otro Administrador, ni a sí mismo', '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
+      });
+      return
+    }
+
+    const id = usuario._id
+
     const { isConfirmed } = await Swal.fire({
-      title: '¿Borrar al Usuario '+ userName +'?',
+      title: '¿Borrar al Usuario '+ usuario.userName +'?',
       showCancelButton: true
     })
 
     if (isConfirmed) {
       this.adminSrv.deleteUser(id)
       .subscribe((res:any) => {
-        Swal.fire(res.msg + ' ' + userName, 'Completado')
+        this.toast.success(res.msg + ' ' + usuario.userName, '', {
+          timeOut: 5000,
+          progressBar: true,
+          progressAnimation: 'decreasing',
+          positionClass: 'toast-top-right',
+        });
         this.cargarUsuarios()
       }, (err)=> {
-          console.warn(err)
-          this.snack.open(err.error.msg, 'Error', {
-          duration: 5000,
-          verticalPosition: 'bottom',
-          horizontalPosition: 'center'
+        console.warn(err) 
+        this.toast.error(err.error.msg, '', {
+          timeOut: 5000,
+          progressBar: true,
+          progressAnimation: 'decreasing',
+          positionClass: 'toast-top-right',
         });
       })
     }
@@ -87,57 +110,100 @@ export class AdministradorComponent implements OnInit {
         Swal.fire(res.msg, 'Completado')
         this.cargarUsuarios()
       }, (err)=> {
-          console.warn(err)
-          this.snack.open(err.error.msg, 'Error', {
-          duration: 5000,
-          verticalPosition: 'bottom',
-          horizontalPosition: 'center'
+        console.warn(err) 
+        this.toast.error(err.error.msg, '', {
+          timeOut: 5000,
+          progressBar: true,
+          progressAnimation: 'decreasing',
+          positionClass: 'toast-top-right',
         });
       })
     }
   }
 
-  changeState(id: string){
+  changeState(usuario: Usuario){
+
+    if (usuario.activo && usuario.tipo == 0 && this.loggedUser != 'ADMIN') {
+      this.toast.error('Un Administrador no puede INACTIVAR a otro Administrador, ni a sí mismo', '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
+      });
+      return
+    }
+
+    const id = usuario._id
+
     this.adminSrv.changeState(id)
     .subscribe((res:any) => {
-      // Swal.fire(res.msg, 'Completado')
-      this.snack.open(res.msg, 'Completado', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+
+      this.toast.success(res.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
+     
       this.cargarUsuarios()
+
     }, (err)=> {
-        console.warn(err)
-        this.snack.open(err.error.msg, 'Error', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+      console.warn(err) 
+      this.toast.error(err.error.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
     })
   }
 
-  unlock(id: string){
+  unlock(usuario: Usuario){
+    if (!usuario.bloqueado && usuario.tipo == 0 && this.loggedUser != 'ADMIN') {
+      this.toast.error('Un Administrador no puede BLOQUEAR a otro Administrador, ni a sí mismo', '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
+      });
+      return
+    }
+
+    const id = usuario._id
+
     this.adminSrv.unlock(id)
     .subscribe((res:any) => {
-      // Swal.fire(res.msg, 'Completado')
-      this.snack.open(res.msg, 'Completado', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+      this.toast.success(res.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
       this.cargarUsuarios()
     }, (err)=> {
-        console.warn(err)
-        this.snack.open(err.error.msg, 'Error', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+      console.warn(err) 
+      this.toast.error(err.error.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
     })
   }
   
-  openDialog(id: string): void {
+  openDialog(usuario: Usuario): void {
+
+    if (usuario.tipo == 0 && this.loggedUser != 'ADMIN') {
+      this.toast.error('Un Administrador no puede EDITAR a otro Administrador, ni a sí mismo', '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
+      });
+      return
+    }
+
+    const id = usuario._id
 
     this.adminSrv.getUser(id)
     .subscribe((res:any) => {
@@ -163,21 +229,20 @@ export class AdministradorComponent implements OnInit {
         this.cargarUsuarios()
       });
     }, (err)=> {
-        console.warn(err)
-        this.snack.open(err.error.msg, 'Error', {
-        duration: 5000,
-        verticalPosition: 'bottom',
-        horizontalPosition: 'center'
+      console.warn(err) 
+      this.toast.error(err.error.msg, '', {
+        timeOut: 5000,
+        progressBar: true,
+        progressAnimation: 'decreasing',
+        positionClass: 'toast-top-right',
       });
     })
   }
 
   hideLoader(){
-
     setTimeout(()=>{
       document.querySelector(".loader")?.classList.add("loader--hidden")
     },1500)
-
   }
 
 }
